Provide env var that can be set instead of config value
Currently a change must be made to the configuration file in keycloak
to decide which directory the metrics data should be stored within.
An environment variable can be specified in the configuration, however
this still requires the configuration file to be modified.

This change will check an environment variable if no config value has
been provided. This allows the location of the metrics to be changed
without modifying any of the files within the original keycloak image.